Refactor Board to ensure one queen per row / col
Instead of verifying we only have one Queen per row or column, change
our data structure so this is implicit - we store just the row num of
the queen in each row. This doesn't gurantee row uniqueness, some sort
of a set would, but there's no nice interface for that in Java.