Board: Use guava for permuts(), cleanup diag check
Use the Guava library for computing perumtations of the board layout,
the interface is a bit wonky, but it works.

Cleanup the diag checking, it didn't include one of the diagonals.
Adding a second HashSet for the second set of diagonals was messy,
switch to using more Streams. Streams everywhere.